+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Register this Doorkeeper extension as `device_code` custom OAuth Grant Flow.
+
 ### Changed
 - Upgrade `doorkeeper` dependency, matching versions `~> 5.5`.
 - Update generated files according to the boilerplate from rails `6.1.3.1`.